Redid Math types in Angel Script to account for alignment

+++ coreutilities/Components.h
@@ -268,7 +268,7 @@ namespace FlexKit
         }
 
         ComponentID ID;
-        uint32_t    componentSize;
+        uint16_t    componentSize;
         union
         {
             ComponentViewBase* _ptr;
